Fixed bug where game data did not save
The issue was that I was saving the score to the wrong ScoreManager variable in GameUi. I was saving it to the actual ScoreManager highscore variable when I should've been using the getter and setter methods to interact with the highscore variable, NOT the highscore variable itself. bad practice to access those global class variables directly i think...?